requirements.txt
aiogram==2.20
APScheduler==3.9.1
python-dotenv==0.21.0
SQLAlchemy==1.4.35
